# Description
This PR closes #562 by speeding up our points plotting. It gives a max number of points to be randomly sampled for thumbnail generation, similar to how there was a maximum number of vectors in the vectors layer.

This speeds up plotting a large number of points considerably, but things are still too slow for ideal usage. Plotting ~100_000 points takes ~20 seconds on my macbook pro, plotting ~1_000_000 points takes ~3 minutes 25 seconds (roughly 10x as long). Not sure what @ambrosejcarr you were seeing before. I havn't done any profiling yet to see where the bottle necks are, but this still feels much slower than what I think vispy can handle.

Once vispy has these points it's pretty fast to pan / zoom, though changing color still appears slow - see here for interacting with 1M points:

![1_M_points](https://user-images.githubusercontent.com/6531703/66020232-47f7d980-e49b-11e9-85a3-87259e62cfa6.gif)

I think in general it's time for us to consider performance benchmarking a little more seriously and I'd like us to set up some infrastructure for doing that more seriously before we start to chase down all our potential performance gains. Some of this is discussed in #476 

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)

# How has this been tested?
<!-- Please describe the tests that you ran to verify your changes. -->
- [ ] example: the test suite for my feature covers cases x, y, and z
- [ ] example: all tests pass with my change

## Final checklist:
- [ ] My PR is the minimum possible work for the desired functionality
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works