@jni 2019-10-02
@sofroniewn I wonder if the cursor movement callbacks are a bottleneck now?