@ambrosejcarr 2019-10-02
@sofroniewn since you're considering benchmarking, it might be worth looping in @ttung who has been thinking about what it would take to do this for starfish.