@ambrosejcarr 2019-10-02
For my current use case, I can probably sub-sample down to ~ 100k points for demonstration purposes, thanks for the quick work @sofroniewn! If it guides additional exploration, eventually I'd hope to find something that could plot <10m points in a few seconds.