@sofroniewn 2019-10-03
Would love to chat with @ttung about benchmarks - will follow up with him. As discussed in #476 going with the practices established in https://scikit-image.org/docs/dev/contribute.html#benchmarks and using https://asv.readthedocs.io/en/stable/ seems like a good place for us to start.

@jni no progress on getting a dedicated machine, but that shouldn't hold us back at this point. Lots of things to do to get going.

I don't think that it's the cursor movement callbacks holding us back as the latency is on loading and once the 1M points are in then its all very fast, but the cursor movement callbacks are still getting called then. I'll dig in a bit more before merging this, if I find anything easy to fix I'll add it, otherwise we can follow up later. This PR should at least make things a little better.