@sofroniewn 2019-10-09
I've merged the benchmarking PR in here - the changes in this PR wrt to the thumbnail generation along cause the following benchmark changes

Before:

```
[ 75.00%] ··· benchmark_points_layer.Points2DSuite.time_create_layer                                                                                                                                      ok
[ 75.00%] ··· ======== ============
               param1
              -------- ------------
                 16     4.27±0.4ms
                 64     9.24±0.2ms
                256      29.1±1ms
                1024     108±3ms
                4096     423±20ms
               16384    1.83±0.2s
               65536    6.92±0.03s
              ======== ============
```

After:
```
[ 25.00%] ··· benchmark_points_layer.Points2DSuite.time_create_layer                                                                                                                                      ok
[ 25.00%] ··· ======== ==========
               param1
              -------- ----------
                 16     5.13±0ms
                 64     9.92±0ms
                256     29.9±0ms
                1024    100±0ms
                4096    124±0ms
               16384    123±0ms
               65536    131±0ms
              ======== ==========
```

Which is a big improvement for basically no functionality loss. We can probably make the thumbnail stuff even better with the kdtree, but we can wait on that for a future PR.

Note the benchmarks are only are the points layer model right now, so don't include the vispy stuff